requirements.txt
anyio==3.7.1
APScheduler==3.9.1.post1
boto3==1.34.112
botocore==1.34.112
cachetools==5.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.1
fonttools==4.51.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.7
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.0a0
pytube==15.0.0
pytz==2024.1
rfc3986==1.5.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tornado==6.4
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
